2021-07-02 20:52 chan tooltip created like_list

+++ ATeam/src/main/java/com/rental/domain/PageDTO.java
@@ -30,6 +30,6 @@ public class PageDTO {
 		}
 		this.prev = this.startPage > 1;
 		this.next = this.endPage < realEnd;
-		this.lastPage = (int)Math.ceil(total / cri.getAmount());
+		this.lastPage = (int)Math.ceil(total /(double) cri.getAmount());
 	}
 }
